Bugs fixed

+++ src/Preprocessing/Convolution.c
@@ -12,9 +12,9 @@ unsigned char **ApplyFilter(Image image, double filter[9])
     unsigned char **tmp_matrix = CreateCharMatrix(size_x, size_y);
 
     //Traversing image.matrix
-    for (size_t x = 1; x < size_y-1; x++)
+    for (size_t x = 1; (int)x < size_y-1; x++)
     {
-        for (size_t y = 1; y < size_x-1; y++)
+        for (size_t y = 1; (int)y < size_x-1; y++)
         {
 
             size_t filter_index = 0;
@@ -23,9 +23,9 @@ unsigned char **ApplyFilter(Image image, double filter[9])
             double divisor = 0;
 
             //Applying matrix
-            for (size_t index_y = y - 1; index_y < y + 2; index_y++)
+            for (int index_y = y - 1; index_y < y + 2; index_y++)
             {
-                for (size_t index_x = x - 1; index_x < x + 2; index_x++)
+                for (int index_x = x - 1; index_x < x + 2; index_x++)
                 {
                     //Skipping particular cases (out of bound, (x,y))
                     if (index_x == -1 || index_x >= size_x || index_y == -1 ||
@@ -88,9 +88,9 @@ void SobelsEdgeDetection(Image image)
     int size_x = image.width;
     int size_y = image.height;
 
-    for (size_t x = 0; x < size_y; x++)
+    for (int x = 0; x < size_y; x++)
     {
-        for (size_t y = 0; y < size_x; y++)
+        for (int y = 0; y < size_x; y++)
         {
             double gradient = sqrt( pow(matrix_gx[y][x], 2)
             + pow(matrix_gy[y][x], 2) );
